Move XCB/GLX Wayland workaround into napari/__init__.py
Applying the workaround from __main__.main() ran too late for
`import napari; napari.Viewer()` scripts: Viewer construction
transitively imports napari._vispy.canvas, which imports
OpenGL.error and triggers OpenGL.platform's one-shot _load().
With WAYLAND_DISPLAY set and PYOPENGL_PLATFORM unset, PyOpenGL
locks into the EGL backend; later env-var writes are ignored.
Qt-on-XCB then paired with PyOpenGL-on-EGL caused the
"no valid context" / "EventEmitter loop detected!" crash on
Linux+Wayland+Nvidia.

Set the env vars at napari import time (matching the
_check_numpy_version.py pattern) so the workaround applies to
both `napari` CLI invocations and `import napari` scripts before
any vispy/PyOpenGL module is loaded.
