+++ src/napari/__main__.py
@@ -5,7 +5,6 @@ napari command line viewer.
 import argparse
 import contextlib
 import logging
-import os
 import sys
 import warnings
 from ast import literal_eval
@@ -371,28 +370,7 @@ def _run_plugin_module(mod, plugin_name):
     run()
 
 
-def _fix_wayland_opengl() -> None:
-    """Set QT_QPA_PLATFORM=xcb and PYOPENGL_PLATFORM=glx on Linux+Wayland.
-
-    On Linux+Wayland with Nvidia hardware, the Nvidia driver's incomplete
-    Wayland support causes napari to crash on startup across Qt bindings.
-    Forcing XCB (X11 via XWayland) + GLX is the recommended workaround. Uses
-    ``os.environ.setdefault`` so any user-set values are never overridden.
-    See https://github.com/napari/napari/issues/8808.
-    """
-    if sys.platform != 'linux':
-        return
-    wayland_active = bool(os.environ.get('WAYLAND_DISPLAY')) or (
-        os.environ.get('XDG_SESSION_TYPE', '').lower() == 'wayland'
-    )
-    if not wayland_active:
-        return
-    os.environ.setdefault('QT_QPA_PLATFORM', 'xcb')
-    os.environ.setdefault('PYOPENGL_PLATFORM', 'glx')
-
-
 def main():
-    _fix_wayland_opengl()
     _run()
